refactor: activity route and component (#2968)
* refactor: activity route and component

* chore: lingui messages

* fix: compose content (#2956)

* feat: come back default path

* fix: type error

* refactor: code style

---------


(cherry picked from commit 4d75c2b8d9b4d0310c58c3e4d97db9a4a483f924)
