+++ src/app/(event)/event/[name]/layout.tsx
