+++ src/hooks/useComeback.ts
@@ -1,18 +1,20 @@
+'use client';
+
 import { useRouter } from 'next/navigation.js';
 import { useCallback } from 'react';
 
 import { PageRoute } from '@/constants/enum.js';
 import { useGlobalState } from '@/store/useGlobalStore.js';
 
-export function useComeBack() {
-    const routeChanged = useGlobalState.use.routeChanged();
+export function useComeBack(path = PageRoute.Home) {
     const router = useRouter();
 
     return useCallback(() => {
+        const routeChanged = useGlobalState.getState().routeChanged;
         if (!routeChanged) {
-            router.push(PageRoute.Home);
+            router.push(path);
             return;
         }
         router.back();
-    }, [routeChanged, router]);
+    }, [path, router]);
 }
